+++ R/add_modifier.R
@@ -29,7 +29,6 @@ add_modifier <- function(phenos,
     messager("Annotating phenos with Modifiers",v=verbose)
     utils::data("hpo_modifiers", package = "HPOExplorer")
     hpo_modifiers <- get("hpo_modifiers")
-    hpo_modifiers[Severity_score_min==Inf,]$Severity_score_min <- NA
 
     #### Add diseases ####
     # DiseaseNames <- hpo$name[hpo$name %in% unique(hpo_modifiers$DiseaseName)]