+++ R/get_gene_lists.R
@@ -15,18 +15,20 @@
 #' @export
 #' @importFrom stats setNames
 #' @examples
-#' phenotypes <- c("Focal motor seizures","HP:0000002","HP:0000003")
+#' phenotypes <- c("Focal motor seizure","HP:0000002","HP:0000003")
 #' gene_lists <- get_gene_lists(phenotypes = phenotypes)
 get_gene_lists <- function(phenotypes,
                            phenotype_to_genes = load_phenotype_to_genes(),
                            hpo = get_hpo(),
                            as_list = FALSE,
                            hpo_id_labels = TRUE) {
+  # templateR:::args2vars(get_gene_lists)
   ID <- NULL;
 
   hpo_ids <- harmonise_phenotypes(phenotypes = phenotypes,
                                   hpo = hpo,
-                                  as_hpo_ids = TRUE)
+                                  as_hpo_ids = TRUE,
+                                  verbose = TRUE)
   p2g <- phenotype_to_genes[ID %in% hpo_ids]
   #### Return at data.table #####
   if(isFALSE(as_list)) return(p2g)