+++ tests/testthat/test-ggnetwork_plot.R
@@ -3,13 +3,13 @@ test_that("ggnetwork_plot works", {
   #### make_phenos_dataframe ####
   phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
   testthat::expect_true(methods::is(phenos,"data.table"))
-  testthat::expect_equal(nrow(phenos),17)
+  testthat::expect_equal(nrow(phenos),23)
 
   #### make_network_object ####
   phenoNet <- make_network_object(phenos = phenos,
                                   colour_var = "ontLvl_geneCount_ratio")
   testthat::expect_true(methods::is(phenoNet,"data.frame"))
-  testthat::expect_equal(nrow(phenoNet),33)
+  testthat::expect_equal(nrow(phenoNet),45)
 
   #### ggnetwork_plot ####
   plt <- ggnetwork_plot(phenoNet = phenoNet,